Rename BufferedSumoLogicAppender to SumoLogicAppender for parity with log4j2 plugin
